bench_surreal_deal/pyproject.toml
[tool.poetry]
name = "bench-surreal-deal"
version = "0.1.0"
description = "internal benchmark"
authors = ["Alexander <post@alexanderfridiksson.com>"]
readme = "README.md"
packages = [{include = "bench_surreal_deal"}]

[tool.poetry.dependencies]
python = "^3.10"
pymongo = "^4.4.1"
neo4j = "^5.11.0"
surrealdb-beta = "^0.0.3"
mimesis = "^11.1.0"
ipykernel = "^6.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
